+++ backend/src/main/java/com/darass/darass/auth/oauth/AuthenticationPrincipalConfig.java
@@ -1,7 +1,7 @@
 package com.darass.darass.auth.oauth;
 
-import com.darass.darass.auth.oauth.controller.AuthenticationPrincipalArgumentResolver;
-import com.darass.darass.auth.oauth.controller.RequiredLoginArgumentResolver;
+import com.darass.darass.auth.oauth.controller.argumentresolver.AuthenticationPrincipalArgumentResolver;
+import com.darass.darass.auth.oauth.controller.argumentresolver.RequiredLoginArgumentResolver;
 import com.darass.darass.auth.oauth.service.OAuthService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
