+++ backend/src/main/java/com/darass/darass/auth/oauth/controller/argumentresolver/RequiredLoginArgumentResolver.java
@@ -1,9 +1,9 @@
-package com.darass.darass.auth.oauth.controller;
+package com.darass.darass.auth.oauth.controller.argumentresolver;
 
-import com.darass.darass.auth.oauth.domain.AuthenticationPrincipal;
+import com.darass.darass.auth.oauth.domain.RequiredLogin;
 import com.darass.darass.auth.oauth.infrastructure.AuthorizationExtractor;
 import com.darass.darass.auth.oauth.service.OAuthService;
-import com.darass.darass.user.domain.GuestUser;
+import com.darass.darass.exception.ExceptionWithMessageAndCode;
 import com.darass.darass.user.domain.User;
 import java.util.Objects;
 import javax.servlet.http.HttpServletRequest;
@@ -15,13 +15,13 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
 @RequiredArgsConstructor
-public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArgumentResolver {
+public class RequiredLoginArgumentResolver implements HandlerMethodArgumentResolver {
 
     private final OAuthService oAuthService;
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(AuthenticationPrincipal.class);
+        return parameter.hasParameterAnnotation(RequiredLogin.class);
     }
 
     @Override
@@ -30,10 +30,9 @@ public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArg
         String accessToken = AuthorizationExtractor
             .extract(Objects.requireNonNull(webRequest.getNativeRequest(HttpServletRequest.class)));
 
-        if (Objects.isNull(accessToken) || accessToken.isEmpty()) {
-            return new GuestUser();
+        if (Objects.isNull(accessToken)) {
+            throw ExceptionWithMessageAndCode.SHOULD_LOGIN.getException();
         }
-
         return oAuthService.findSocialLoginUserByAccessToken(accessToken);
     }
 }
